modularization: move PokemonEndpoint to a separate file
